+++ src/main/java/com/rudolph/Weevo/chat/dto/summary/ChatRoomSummary.java
@@ -3,7 +3,7 @@ package com.rudolph.Weevo.chat.dto.summary;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.rudolph.Weevo.chat.domain.Chat;
 import com.rudolph.Weevo.chat.domain.ChatRoom;
-import com.rudolph.Weevo.member.domain.Member;
+import com.rudolph.Weevo.memberr.domain.Member;
 import lombok.Builder;
 import lombok.Getter;
