+++ src/main/java/com/rudolph/Weevo/chat/service/ChatRoomService.java
@@ -15,8 +15,8 @@ import com.rudolph.Weevo.course.domain.Course;
 import com.rudolph.Weevo.course.service.CourseService;
 import com.rudolph.Weevo.global.common.code.ErrorStatus;
 import com.rudolph.Weevo.global.exception.GeneralException;
-import com.rudolph.Weevo.member.domain.Member;
-import com.rudolph.Weevo.member.service.MemberService;
+import com.rudolph.Weevo.memberr.domain.Member;
+import com.rudolph.Weevo.memberr.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
