+++ src/main/java/com/rudolph/Weevo/course/controller/CourseController.java
@@ -108,13 +108,6 @@ public class CourseController {
         MyCoursesResponse resp = courseService.listMyCourses(user.getMemberId());
         return ApiResponse.onSuccess(SuccessStatus._OK, resp);
     }
-  
-  
-//     @GetMapping("/my-course")   //내 강의 조회
-//     public ResponseEntity<MyCourseListDto> getMyCourses(@AuthenticationPrincipal CustomUserPrincipal principal) {
-//         MyCourseListDto myCourseList = courseService.getMyCourseList(principal);
-//         return ResponseEntity.ok(myCourseList);
-//     }
 
 
 }
