+++ src/main/java/com/rudolph/Weevo/course/domain/Course.java
@@ -3,7 +3,7 @@ package com.rudolph.Weevo.course.domain;
 
 import com.rudolph.Weevo.course.domain.enums.CourseCity;
 import com.rudolph.Weevo.global.domain.BaseEntity;
-import com.rudolph.Weevo.member.domain.Member;
+import com.rudolph.Weevo.memberr.domain.Member;
 import com.rudolph.Weevo.course.domain.enums.CourseCategory;
 import com.rudolph.Weevo.course.domain.enums.CourseType;
 import jakarta.persistence.*;
