+++ src/main/java/com/rudolph/Weevo/course/service/CourseService.java
@@ -5,14 +5,12 @@ import com.rudolph.Weevo.course.dto.request.CreateCourseRequest;
 import com.rudolph.Weevo.course.dto.response.*;
 import com.rudolph.Weevo.course.domain.*;
 import com.rudolph.Weevo.course.repository.CourseBookmarkRepository;
-import com.rudolph.Weevo.auth.security.CustomUserPrincipal;
 import com.rudolph.Weevo.course.domain.Course;
 import com.rudolph.Weevo.course.domain.MemberCourse;
-import com.rudolph.Weevo.course.dto.response.MyCourseListDto;
 import com.rudolph.Weevo.course.repository.CourseRepository;
 import com.rudolph.Weevo.course.repository.CourseSpecification;
-import com.rudolph.Weevo.member.domain.Member;
-import com.rudolph.Weevo.member.repository.MemberRepository;
+import com.rudolph.Weevo.memberr.domain.Member;
+import com.rudolph.Weevo.memberr.repository.MemberRepository;
 import com.rudolph.Weevo.global.common.code.ErrorStatus;
 import com.rudolph.Weevo.global.exception.GeneralException;
 import com.rudolph.Weevo.global.service.S3Service;
@@ -30,7 +28,6 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 import java.util.List;
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -235,18 +232,6 @@ public class CourseService {
                 .teachingCourses(teaching)
                 .enrolledCourses(enrolled)
                 .build();
-    }
-  
-//     @Transactional(readOnly = true)     //내 강의 조회
-//     public MyCourseListDto getMyCourseList(CustomUserPrincipal principal) {
-//         UUID memberId = principal.getMemberId();
-//         Member member = memberRepository.findByMemberId(memberId)
-//                 .orElseThrow(() -> new GeneralException(ErrorStatus.MEMBER_NOT_FOUND));
-//         List<MemberCourse> memberCourses = memberCourseRepository.findAllByMember(member);
-//         List<Course> myCourses = memberCourses.stream()
-//                 .map(MemberCourse::getCourse)
-//                 .toList();
-//         return MyCourseListDto.from(myCourses);
     }
 
 }
