+++ src/main/java/com/rudolph/Weevo/memberr/controller/MemberController.java
@@ -1,14 +1,14 @@
-package com.rudolph.Weevo.member.controller;
+package com.rudolph.Weevo.memberr.controller;
 
 import com.rudolph.Weevo.auth.security.CustomUserPrincipal;
-import com.rudolph.Weevo.member.dto.response.MemberDetailResponse;
-import com.rudolph.Weevo.member.dto.response.MemberListResponse;
-import com.rudolph.Weevo.member.repository.MemberRepository;
-import com.rudolph.Weevo.member.service.MemberService;
+import com.rudolph.Weevo.memberr.dto.response.MemberDetailResponse;
+import com.rudolph.Weevo.memberr.dto.response.MemberListResponse;
+import com.rudolph.Weevo.memberr.repository.MemberRepository;
+import com.rudolph.Weevo.memberr.service.MemberService;
 import com.rudolph.Weevo.global.common.code.ErrorStatus;
 import com.rudolph.Weevo.global.exception.GeneralException;
-import com.rudolph.Weevo.member.domain.Member;
-import com.rudolph.Weevo.member.dto.request.InfoRequest;
+import com.rudolph.Weevo.memberr.domain.Member;
+import com.rudolph.Weevo.memberr.dto.request.InfoRequest;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
