+++ src/main/java/com/rudolph/Weevo/memberr/domain/MemberTalentTag.java
@@ -1,4 +1,4 @@
-package com.rudolph.Weevo.member.domain;
+package com.rudolph.Weevo.memberr.domain;
 
 import com.rudolph.Weevo.tag.domain.Tag;
 import jakarta.persistence.*;
@@ -6,10 +6,10 @@ import lombok.*;
 
 @Entity
 @Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
-public class MemberInterestTag {
+public class MemberTalentTag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
