+++ src/main/java/com/rudolph/Weevo/memberr/dto/request/FixProfileRequestDto.java
@@ -1,4 +1,4 @@
-package com.rudolph.Weevo.member.dto.request;
+package com.rudolph.Weevo.memberr.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
