+++ src/main/java/com/rudolph/Weevo/memberr/dto/request/UpdateInterestTagRequestDto.java
@@ -1,4 +1,4 @@
-package com.rudolph.Weevo.member.dto.request;
+package com.rudolph.Weevo.memberr.dto.request;
 
 import lombok.Getter;
