+++ src/main/java/com/rudolph/Weevo/memberr/dto/response/MemberInterestTagDto.java
@@ -1,17 +1,16 @@
-package com.rudolph.Weevo.member.dto.response;
+package com.rudolph.Weevo.memberr.dto.response;
 
-import com.rudolph.Weevo.member.domain.Member;
-import com.rudolph.Weevo.member.domain.MemberInterestTag;
+import com.rudolph.Weevo.memberr.domain.Member;
+import com.rudolph.Weevo.memberr.domain.MemberInterestTag;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 import java.util.List;
-import java.util.UUID;
 
 @Getter
 @AllArgsConstructor
 public class MemberInterestTagDto {
-    private UUID memberId;
+    private Long memberId;
     private List<String> tagNames;
 
     public static MemberInterestTagDto from(Member member, List<MemberInterestTag> interestTags) {
@@ -19,6 +18,6 @@ public class MemberInterestTagDto {
                 .map(interestTag -> interestTag.getTag().getName())
                 .toList();
 
-        return new MemberInterestTagDto(member.getMemberId(), tagNames);
+        return new MemberInterestTagDto(member.getId(), tagNames);
     }
 }
