+++ src/main/java/com/rudolph/Weevo/memberr/repository/MemberRepositoryCustom.java
@@ -1,6 +1,6 @@
-package com.rudolph.Weevo.member.repository;
+package com.rudolph.Weevo.memberr.repository;
 
-import com.rudolph.Weevo.member.dto.response.MemberListResponse;
+import com.rudolph.Weevo.memberr.dto.response.MemberListResponse;
 import org.springframework.data.domain.Sort;
 
 import java.util.List;
