+++ src/main/java/com/rudolph/Weevo/memberr/repository/MemberRepositoryImpl.java
@@ -1,7 +1,7 @@
-package com.rudolph.Weevo.member.repository;
+package com.rudolph.Weevo.memberr.repository;
 
-import com.rudolph.Weevo.member.domain.Member;
-import com.rudolph.Weevo.member.dto.response.MemberListResponse;
+import com.rudolph.Weevo.memberr.domain.Member;
+import com.rudolph.Weevo.memberr.dto.response.MemberListResponse;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
